refactor: clean up thermodata subpackage
