+++ atmodeller/thermodata/data/nasa_glenn_coefficients.txt
@@ -1,9 +1,9 @@
-# Data from McBride et al. (2002), NASA Glenn Coefficients for Calculating Thermodynamic Properties of Individual Species
+# Data from McBride et al. (2002), NASA Glenn Coefficients for Calculating Thermodynamic Properties of Individual Species, No. 2002-211556 NASA
+# https://ntrs.nasa.gov/citations/20020085330
 # The Hill system is adopted
-# -------------
-# --- Gases ---
-# -------------
+# Header line is below
 Formula State T_min T_max a1 a2 a3 a4 a5 a6 a7 b1 b2
+# --- Gases ---
 Ar g 200 1000 0.0 0.0 2.5 0.0 0.0 0.0 0.0 -7.453750000e02 4.379674910e00
 Ar g 1000 6000 2.010538475e01 -5.992661070e-02 2.500069401e00 -3.992141160e-08 1.205272140e-11 -1.819015576e-15 1.078576636e-19 -7.449939610e02 4.379180110e00
 Ar g 6000 20000 -9.951265080e08 6.458887260e05 -1.675894697e02 2.319933363e-02 -1.721080911e-06 6.531938460e-11 -9.740147729e-16 -5.078300340e06 1.465298484e03
@@ -93,9 +93,7 @@ Si g 6000 20000 -9.286548940e08 5.443989890e05 -1.206739736e02 1.359662698e-02 -
 Xe g 200 1000 0.0 0.0 2.5 0.0 0.0 0.0 0.0 -7.453750000e02 6.164454205e00
 Xe g 1000 6000 4.025226680e03 -1.209507521e01 2.514153347e00 -8.248102080e-06 2.530232618e-09 -3.892333230e-13 2.360439138e-17 -6.685800730e02 6.063710715e00
 Xe g 6000 20000 2.540397456e08 -1.105373774e05 1.382644099e01 1.500614606e-03 -3.935359030e-07 2.765790584e-11 -5.943990574e-16 9.285443830e05 -1.109834556e02
-# -----------------
 # --- Condensed ---
-# -----------------
 C cr 200 600 1.132856760e5 -1.980421677e3 1.365384188e1 -4.636096440e-2 1.021333011e-4 -1.082893179e-7 4.472258860e-11 8.943859760e3 -7.295824740e1
 C cr 600 2000 3.356004410e5 -2.596528368e3 6.948841910 -3.484836090e-3 1.844192445e-6 -5.055205960e-10 5.750639010e-14 1.398412456e4 -4.477183040e1
 C cr 2000 6000 2.023105106e5 -1.138235908e3 3.700279500 -1.833807727e-4 6.343683250e-8 -7.068589480e-12 3.335435980e-16 5.848134850e3 -2.350925275e1
